backend/requirements.txt
flask
flask-cors
requests
pyjwt
random-word
RandomWords
pillow

